+++ src/shared/qrcode/generator.js
@@ -1,0 +1,459 @@
+/**
+ * @copyright Kazuhiko Arase
+ * @copyright davidshimjs
+ * @license MIT
+ * The word "QR Code" is registered trademark of DENSO WAVE INCORPORATED
+ */
+import { QRErrorCorrectLevelMap, QRMode, RS_BLOCK_TABLE } from "./constants";
+import QRPolynomial from "./polynomial";
+import { QRUtil, getTypeNumber } from "./utils";
+
+export class QRCodeModel {
+    /**
+     * @param {String} text
+     * @param {import('./types').ErrorCorrectLevel} errorCorrectLevel
+     */
+    constructor(text, errorCorrectLevel) {
+        this.text = text;
+        this.errorCorrectLevel = errorCorrectLevel;
+        this.typeNumber = getTypeNumber(this.text, this.errorCorrectLevel);
+        this.modules = null;
+        this.moduleCount = 0;
+        this.dataCache = null;
+        this.dataList = [new QR8bitByte(this.text)];
+    }
+
+    static get PAD0() {
+        return 0xec;
+    }
+
+    static get PAD1() {
+        return 0x11;
+    }
+
+    /**
+     * @param {number} row
+     * @param {number} col
+     */
+    isDark(row, col) {
+        if (row < 0 || this.moduleCount <= row || col < 0 || this.moduleCount <= col) {
+            throw new Error(row + "," + col);
+        }
+        return this.modules[row][col];
+    }
+
+    getModuleCount() {
+        return this.moduleCount;
+    }
+
+    make() {
+        this.makeImpl(false, this.getBestMaskPattern());
+    }
+
+    /**
+     * @param {boolean} test
+     * @param {Number} maskPattern
+     */
+    makeImpl(test, maskPattern) {
+        this.moduleCount = this.typeNumber * 4 + 17;
+        this.modules = new Array(this.moduleCount);
+        for (let row = 0; row < this.moduleCount; row++) {
+            this.modules[row] = new Array(this.moduleCount);
+            for (let col = 0; col < this.moduleCount; col++) {
+                this.modules[row][col] = null;
+            }
+        }
+        this.setupPositionProbePattern(0, 0);
+        this.setupPositionProbePattern(this.moduleCount - 7, 0);
+        this.setupPositionProbePattern(0, this.moduleCount - 7);
+        this.setupPositionAdjustPattern();
+        this.setupTimingPattern();
+        this.setupTypeInfo(test, maskPattern);
+        if (this.typeNumber >= 7) {
+            this.setupTypeNumber(test);
+        }
+        if (this.dataCache == null) {
+            this.dataCache = QRCodeModel.createData(this.typeNumber, this.errorCorrectLevel, this.dataList);
+        }
+        this.mapData(this.dataCache, maskPattern);
+    }
+
+    /**
+     * @param {number} row
+     * @param {number} col
+     */
+    setupPositionProbePattern(row, col) {
+        for (let r = -1; r <= 7; r++) {
+            if (row + r <= -1 || this.moduleCount <= row + r) continue;
+            for (let c = -1; c <= 7; c++) {
+                if (col + c <= -1 || this.moduleCount <= col + c) continue;
+                if (
+                    (0 <= r && r <= 6 && (c == 0 || c == 6)) ||
+                    (0 <= c && c <= 6 && (r == 0 || r == 6)) ||
+                    (2 <= r && r <= 4 && 2 <= c && c <= 4)
+                ) {
+                    this.modules[row + r][col + c] = true;
+                } else {
+                    this.modules[row + r][col + c] = false;
+                }
+            }
+        }
+    }
+
+    /**
+     * @returns {Number}
+     */
+    getBestMaskPattern() {
+        let minLostPoint = 0;
+        let pattern = 0;
+        for (let i = 0; i < 8; i++) {
+            this.makeImpl(true, i);
+            let lostPoint = QRUtil.getLostPoint(this);
+            if (i == 0 || minLostPoint > lostPoint) {
+                minLostPoint = lostPoint;
+                pattern = i;
+            }
+        }
+        return pattern;
+    }
+
+    setupTimingPattern() {
+        for (let r = 8; r < this.moduleCount - 8; r++) {
+            if (this.modules[r][6] != null) {
+                continue;
+            }
+            this.modules[r][6] = r % 2 == 0;
+        }
+        for (let c = 8; c < this.moduleCount - 8; c++) {
+            if (this.modules[6][c] != null) {
+                continue;
+            }
+            this.modules[6][c] = c % 2 == 0;
+        }
+    }
+
+    setupPositionAdjustPattern() {
+        let pos = QRUtil.getPatternPosition(this.typeNumber);
+        for (let i = 0; i < pos.length; i++) {
+            for (let j = 0; j < pos.length; j++) {
+                let row = pos[i];
+                let col = pos[j];
+                if (this.modules[row][col] != null) {
+                    continue;
+                }
+                for (let r = -2; r <= 2; r++) {
+                    for (let c = -2; c <= 2; c++) {
+                        if (r == -2 || r == 2 || c == -2 || c == 2 || (r == 0 && c == 0)) {
+                            this.modules[row + r][col + c] = true;
+                        } else {
+                            this.modules[row + r][col + c] = false;
+                        }
+                    }
+                }
+            }
+        }
+    }
+
+    /**
+     * @param {boolean} test
+     */
+    setupTypeNumber(test) {
+        let bits = QRUtil.getBCHTypeNumber(this.typeNumber);
+        for (let i = 0; i < 18; i++) {
+            let mod = !test && ((bits >> i) & 1) == 1;
+            this.modules[Math.floor(i / 3)][(i % 3) + this.moduleCount - 8 - 3] = mod;
+        }
+        for (let i = 0; i < 18; i++) {
+            let mod = !test && ((bits >> i) & 1) == 1;
+            this.modules[(i % 3) + this.moduleCount - 8 - 3][Math.floor(i / 3)] = mod;
+        }
+    }
+
+    /**
+     * @param {boolean} test
+     * @param {Number} maskPattern
+     */
+    setupTypeInfo(test, maskPattern) {
+        const data = (this.errorCorrectLevel << 3) | maskPattern;
+        const bits = QRUtil.getBCHTypeInfo(data);
+        for (let i = 0; i < 15; i++) {
+            let mod = !test && ((bits >> i) & 1) == 1;
+            if (i < 6) {
+                this.modules[i][8] = mod;
+            } else if (i < 8) {
+                this.modules[i + 1][8] = mod;
+            } else {
+                this.modules[this.moduleCount - 15 + i][8] = mod;
+            }
+        }
+        for (let i = 0; i < 15; i++) {
+            let mod = !test && ((bits >> i) & 1) == 1;
+            if (i < 8) {
+                this.modules[8][this.moduleCount - i - 1] = mod;
+            } else if (i < 9) {
+                this.modules[8][15 - i - 1 + 1] = mod;
+            } else {
+                this.modules[8][15 - i - 1] = mod;
+            }
+        }
+        this.modules[this.moduleCount - 8][8] = !test;
+    }
+
+    mapData(data, maskPattern) {
+        let inc = -1;
+        let row = this.moduleCount - 1;
+        let bitIndex = 7;
+        let byteIndex = 0;
+        for (let col = this.moduleCount - 1; col > 0; col -= 2) {
+            if (col == 6) col--;
+            while (true) {
+                for (let c = 0; c < 2; c++) {
+                    if (this.modules[row][col - c] == null) {
+                        let dark = false;
+                        if (byteIndex < data.length) {
+                            dark = ((data[byteIndex] >>> bitIndex) & 1) == 1;
+                        }
+                        let mask = QRUtil.getMask(maskPattern, row, col - c);
+                        if (mask) {
+                            dark = !dark;
+                        }
+                        this.modules[row][col - c] = dark;
+                        bitIndex--;
+                        if (bitIndex == -1) {
+                            byteIndex++;
+                            bitIndex = 7;
+                        }
+                    }
+                }
+                row += inc;
+                if (row < 0 || this.moduleCount <= row) {
+                    row -= inc;
+                    inc = -inc;
+                    break;
+                }
+            }
+        }
+    }
+
+    static createData(typeNumber, errorCorrectLevel, dataList) {
+        let rsBlocks = QRRSBlock.getRSBlocks(typeNumber, errorCorrectLevel);
+        const buffer = new QRBitBuffer();
+        for (let i = 0; i < dataList.length; i++) {
+            let data = dataList[i];
+            buffer.put(data.mode, 4);
+            buffer.put(data.getLength(), QRUtil.getLengthInBits(data.mode, typeNumber));
+            data.write(buffer);
+        }
+        let totalDataCount = 0;
+        for (let i = 0; i < rsBlocks.length; i++) {
+            totalDataCount += rsBlocks[i].dataCount;
+        }
+        if (buffer.getLengthInBits() > totalDataCount * 8) {
+            throw new Error("code length overflow. (" + buffer.getLengthInBits() + ">" + totalDataCount * 8 + ")");
+        }
+        if (buffer.getLengthInBits() + 4 <= totalDataCount * 8) {
+            buffer.put(0, 4);
+        }
+        while (buffer.getLengthInBits() % 8 != 0) {
+            buffer.putBit(false);
+        }
+        while (true) {
+            if (buffer.getLengthInBits() >= totalDataCount * 8) {
+                break;
+            }
+            buffer.put(QRCodeModel.PAD0, 8);
+            if (buffer.getLengthInBits() >= totalDataCount * 8) {
+                break;
+            }
+            buffer.put(QRCodeModel.PAD1, 8);
+        }
+        return QRCodeModel.createBytes(buffer, rsBlocks);
+    }
+
+    static createBytes(buffer, rsBlocks) {
+        let offset = 0;
+        let maxDcCount = 0;
+        let maxEcCount = 0;
+        let dcdata = new Array(rsBlocks.length);
+        let ecdata = new Array(rsBlocks.length);
+        for (let r = 0; r < rsBlocks.length; r++) {
+            let dcCount = rsBlocks[r].dataCount;
+            let ecCount = rsBlocks[r].totalCount - dcCount;
+            maxDcCount = Math.max(maxDcCount, dcCount);
+            maxEcCount = Math.max(maxEcCount, ecCount);
+            dcdata[r] = new Array(dcCount);
+            for (let i = 0; i < dcdata[r].length; i++) {
+                dcdata[r][i] = 0xff & buffer.buffer[i + offset];
+            }
+            offset += dcCount;
+            let rsPoly = QRUtil.getErrorCorrectPolynomial(ecCount);
+            let rawPoly = new QRPolynomial(dcdata[r], rsPoly.getLength() - 1);
+            let modPoly = rawPoly.mod(rsPoly);
+            ecdata[r] = new Array(rsPoly.getLength() - 1);
+            for (let i = 0; i < ecdata[r].length; i++) {
+                let modIndex = i + modPoly.getLength() - ecdata[r].length;
+                ecdata[r][i] = modIndex >= 0 ? modPoly.get(modIndex) : 0;
+            }
+        }
+        let totalCodeCount = 0;
+        for (let i = 0; i < rsBlocks.length; i++) {
+            totalCodeCount += rsBlocks[i].totalCount;
+        }
+        let data = new Array(totalCodeCount);
+        let index = 0;
+        for (let i = 0; i < maxDcCount; i++) {
+            for (let r = 0; r < rsBlocks.length; r++) {
+                if (i < dcdata[r].length) {
+                    data[index++] = dcdata[r][i];
+                }
+            }
+        }
+        for (let i = 0; i < maxEcCount; i++) {
+            for (let r = 0; r < rsBlocks.length; r++) {
+                if (i < ecdata[r].length) {
+                    data[index++] = ecdata[r][i];
+                }
+            }
+        }
+        return data;
+    }
+}
+
+class QR8bitByte {
+    /**
+     * @param {string} data
+     */
+    constructor(data) {
+        this.mode = QRMode.MODE_8BIT_BYTE;
+        this.data = data;
+        this.parsedData = [];
+
+        // Added to support UTF-8 Characters
+        for (let i = 0, l = this.data.length; i < l; i++) {
+            let byteArray = [];
+            let code = this.data.charCodeAt(i);
+
+            if (code > 0x10000) {
+                byteArray[0] = 0xf0 | ((code & 0x1c0000) >>> 18);
+                byteArray[1] = 0x80 | ((code & 0x3f000) >>> 12);
+                byteArray[2] = 0x80 | ((code & 0xfc0) >>> 6);
+                byteArray[3] = 0x80 | (code & 0x3f);
+            } else if (code > 0x800) {
+                byteArray[0] = 0xe0 | ((code & 0xf000) >>> 12);
+                byteArray[1] = 0x80 | ((code & 0xfc0) >>> 6);
+                byteArray[2] = 0x80 | (code & 0x3f);
+            } else if (code > 0x80) {
+                byteArray[0] = 0xc0 | ((code & 0x7c0) >>> 6);
+                byteArray[1] = 0x80 | (code & 0x3f);
+            } else {
+                byteArray[0] = code;
+            }
+
+            this.parsedData.push(byteArray);
+        }
+
+        this.parsedData = Array.prototype.concat.apply([], this.parsedData);
+
+        if (this.parsedData.length != this.data.length) {
+            this.parsedData.unshift(191);
+            this.parsedData.unshift(187);
+            this.parsedData.unshift(239);
+        }
+    }
+
+    getLength() {
+        return this.parsedData.length;
+    }
+
+    /**
+     * @param {QRBitBuffer} buffer
+     */
+    write(buffer) {
+        for (let i = 0, l = this.parsedData.length; i < l; i++) {
+            buffer.put(this.parsedData[i], 8);
+        }
+    }
+}
+
+class QRRSBlock {
+    constructor(totalCount, dataCount) {
+        this.totalCount = totalCount;
+        this.dataCount = dataCount;
+    }
+
+    static getRSBlocks(typeNumber, errorCorrectLevel) {
+        let rsBlock = QRRSBlock.getRsBlockTable(typeNumber, errorCorrectLevel);
+        if (rsBlock == undefined) {
+            throw new Error("bad rs block @ typeNumber:" + typeNumber + "/errorCorrectLevel:" + errorCorrectLevel);
+        }
+        let length = rsBlock.length / 3;
+        let list = [];
+        for (let i = 0; i < length; i++) {
+            let count = rsBlock[i * 3 + 0];
+            let totalCount = rsBlock[i * 3 + 1];
+            let dataCount = rsBlock[i * 3 + 2];
+            for (let j = 0; j < count; j++) {
+                list.push(new QRRSBlock(totalCount, dataCount));
+            }
+        }
+        return list;
+    }
+
+    static getRsBlockTable(typeNumber, errorCorrectLevel) {
+        switch (errorCorrectLevel) {
+            case QRErrorCorrectLevelMap.L:
+                return RS_BLOCK_TABLE[(typeNumber - 1) * 4 + 0];
+            case QRErrorCorrectLevelMap.M:
+                return RS_BLOCK_TABLE[(typeNumber - 1) * 4 + 1];
+            case QRErrorCorrectLevelMap.Q:
+                return RS_BLOCK_TABLE[(typeNumber - 1) * 4 + 2];
+            case QRErrorCorrectLevelMap.H:
+                return RS_BLOCK_TABLE[(typeNumber - 1) * 4 + 3];
+            default:
+                return undefined;
+        }
+    }
+}
+
+class QRBitBuffer {
+    constructor() {
+        this.buffer = [];
+        this.length = 0;
+    }
+
+    /**
+     * @param {Number} index
+     */
+    get(index) {
+        const bufIndex = Math.floor(index / 8);
+        return ((this.buffer[bufIndex] >>> (7 - (index % 8))) & 1) == 1;
+    }
+
+    /**
+     * @param {Number} num
+     * @param {Number} length
+     */
+    put(num, length) {
+        for (let i = 0; i < length; i++) {
+            this.putBit(((num >>> (length - i - 1)) & 1) == 1);
+        }
+    }
+
+    getLengthInBits() {
+        return this.length;
+    }
+
+    /**
+     * @param {Boolean} bit
+     */
+    putBit(bit) {
+        let bufIndex = Math.floor(this.length / 8);
+        if (this.buffer.length <= bufIndex) {
+            this.buffer.push(0);
+        }
+        if (bit) {
+            this.buffer[bufIndex] |= 0x80 >>> this.length % 8;
+        }
+        this.length++;
+    }
+}