+++ src/types/plugins/omemo/templates/profile.d.ts
@@ -1,3 +1,3 @@
-declare function _default(el: any): import("lit").TemplateResult<1>;
+declare function _default(el: import("../profile").Profile): import("lit").TemplateResult<1>;
 export default _default;
 //# sourceMappingURL=profile.d.ts.map